[EXSHAPP-0573]: [Bug] SubunitValidationService uses Double arithmetic, outdated Cloud Function comment, and permissive Firestore rules (#584)

+++ data/local/src/main/kotlin/es/pedrazamiguez/expenseshareapp/data/local/entity/SubunitEntity.kt
@@ -4,6 +4,7 @@ import androidx.room.Entity
 import androidx.room.ForeignKey
 import androidx.room.Index
 import androidx.room.PrimaryKey
+import java.math.BigDecimal
 
 @Entity(
     tableName = "subunits",
@@ -23,7 +24,7 @@ data class SubunitEntity(
     val groupId: String,
     val name: String,
     val memberIds: List<String>,
-    val memberShares: Map<String, Double>,
+    val memberShares: Map<String, BigDecimal>,
     val createdBy: String,
     val createdAtMillis: Long?,
     val lastUpdatedAtMillis: Long?
